requirements.txt
Flask==3.1.3
flask_cors==6.0.2
langchain_chroma==1.1.0
langchain_community==0.4.1
langchain_core==1.2.22
protobuf==7.34.1
python-dotenv==1.2.2
Requests==2.32.5
